Deep copy object before saving
It looks like node-gcloud modifies Objects in some scenarios when
saving them to the datastore. The case I was able to reproduce is
it modifying arrays to be of type 'arrayValue' when the array is
inside another class. This case is represented in the included
test case.

On our current state, a shallow copy of the object would be enough,
but since the problem could arise again if we have the class
instance with an array deeper in the object structure. A test case
has also been added to cover this case.